# Improve '/page/product_brands' search domain