@FrancescoBallerini 2024-03-26
> What about making the brand `website_published` computed stored depending on the products `is_pusblished` state?

Thanks for the review! I would keep the brandh `website_published` independent from the products `is_published` state: a user might need to temporarily unpublish a brand which has published products. Do you agree?

Thanks