@chienandalu 2024-03-27
This would not be a problem. A computed stored field behaves very much like an onchange. However, it's true that you should decide whether or not to publish your brand depending on the state of the product...

I also wonder if with two separate queries we could gain some performance:

- First one: a read_group for published products grouped by brand.
- Second one: brands published and filtered by those ids gathered in the first query