@FrancescoBallerini 2024-03-27
> I also wonder if with two separate queries we could gain some performance:
> 
> * First one: a read_group for published products grouped by brand.
> * Second one: brands published and filtered by those ids gathered in the first query

Hi @chienandalu I have made some tests, I had never used `read_group` method before so I am not sure if I executed it correctly, however from my tests it seems that the ORM mapping is winning over dobule query with groubpy.

I have prepared a patched version of the module where the controller main.py file is patched so it execute both queries and log execution time, so anyone could make additional tests if interested

[website_sale_product_brand_test_performance.zip](https://github.com/OCA/e-commerce/files/14781069/website_sale_product_brand_test_performance.zip)

From my tests I got similar pattern for every execution:

```
ORM search domain
Time taken: 0.00825832999998965533450245857238769531250000000000 seconds

ORM Groupby
Time taken: 0.01319241300006979145109653472900390625000000000000 seconds
```