@FrancescoBallerini 2024-04-04
@chienandalu I realized that the "time taken" I reported in the last post for the single query (after adding `("is_published", "=", True)` in the domain) was a bit lower than the average I reported in the first post.. I hope that it wasn't misleading, I tested from the same machine but it was probably doing less work in background while testing :-) 

Basically the "time diff" average was about 0,005 seconds between two queries, which is not much but I would keep the search() at this point as it probably improves a bit the code readability as well. 

But let me know if you agree, also if you prefer I could repeat the test and post more lines